requirements.txt
flask
flask_restful
psycopg2
sqlalchemy
ConfigParser
StringGenerator
pyjwt
requests